Remove appModeService, integrate logic into ProblemService
