+++ frontend/src/app/annotate/annotation-input/knowledge-base-form/knowledge-base-form.component.html
@@ -1,5 +1,5 @@
 @let kbControls = form().controls.kbItems.controls;
-@let viewMode = viewMode$ | async;
+@let appMode = appMode$ | async;
 
 <form [formGroup]="form()">
     <fieldset>
@@ -7,7 +7,7 @@
             class="form-label h6 fw-bold d-flex align-items-center justify-content-between"
         >
             <span i18n>Knowledge base items</span>
-            @if (viewMode === 'add' || viewMode === 'edit') {
+            @if (appMode === 'add' || appMode === 'edit') {
             <button
                 type="button"
                 class="btn btn-sm btn-outline-primary"
@@ -28,7 +28,7 @@
     <ul class="ps-0 list-unstyled" formArrayName="kbItems">
         @for (kbItem of kbControls; let i = $index; track `${i}-${kbItem.value.entity1}-${kbItem.value.relationship}-${kbItem.value.entity2}`) {
         <li class="mb-3">
-            @if (viewMode === 'add' || viewMode === 'edit') {
+            @if (appMode === 'add' || appMode === 'edit') {
             <div [formGroupName]="i" class="input-group">
                 <input
                     type="text"
