+++ frontend/src/app/annotate/annotation-input/premises-form/premises-form.component.html
@@ -1,5 +1,5 @@
 @let premiseControls = form().controls.premises.controls;
-@let viewMode = viewMode$ | async;
+@let appMode = appMode$ | async;
 
 <form [formGroup]="form()">
     <fieldset>
@@ -7,7 +7,7 @@
             class="form-label h6 fw-bold d-flex align-items-center justify-content-between"
         >
             <span i18n>Premises</span>
-            @if (viewMode === 'add' || viewMode === 'edit') {
+            @if (appMode === 'add' || appMode === 'edit') {
             <button
                 type="button"
                 class="btn btn-sm btn-outline-primary"
@@ -32,13 +32,13 @@
                         [id]="'premise-' + i"
                         type="text"
                         [class]="
-                            viewMode === 'add' || viewMode === 'edit'
+                            appMode === 'add' || appMode === 'edit'
                                 ? 'form-control'
                                 : 'form-control-plaintext'
                         "
                         [formControlName]="i"
                     />
-                    @if (viewMode === 'add' || viewMode === 'edit') {
+                    @if (appMode === 'add' || appMode === 'edit') {
                     <button
                         type="button"
                         class="btn btn-danger"
@@ -76,7 +76,7 @@
             id="hypothesis"
             type="text"
             [class]="
-                viewMode === 'add' || viewMode === 'edit'
+                appMode === 'add' || appMode === 'edit'
                     ? 'form-control'
                     : 'form-control-plaintext'
             "
